feat: Resolve the 5.7.2 version self-test problem
解决5.7.2版本自测问题

Log: 1. 解决产线审核模式下双盘全盘加密数据盘挂载超时问题
     2. 解决产线审核模式/home/test目录属组导致的登录问题
     3. 解决产线审核模式和保留用户数据的需求兼容问题
     4. 解决10die u模块开关问题
     5. 解决保留用户数据安装失败问题
     6. 解决双盘全盘加密时，修改磁盘密码的接口问题

Task: https://pms.uniontech.com/zentao/task-view-86846.html

+++ check_hooks/check_stop.sh
@@ -3,10 +3,12 @@
 . "/deepin-installer/basic_utils.sh"
 
 clean_disk_cryption() {
-    local key_file=/boot/keyfile
+    local crypt_initramfs=$workspace/etc/cryptsetup-initramfs/conf-hook
     local cryp_conf_file=/etc/crypttab
+    local key_file=/etc/deepin/crypt_keyfile_*.key
 
-    mv ${cryp_conf_file}.real ${cryp_conf_file}
+    [ -f ${crypt_initramfs}.real ] && mv ${crypt_initramfs}.real ${crypt_initramfs}
+    [ -f ${cryp_conf_file}.real ] && mv ${cryp_conf_file}.real ${cryp_conf_file}
     rm -fr $key_file
     /usr/sbin/update-initramfs -u
 }
@@ -18,6 +20,6 @@ clean_check_mode() {
     installer_set "system_check_mode" "false"
 }
 
-clean_disk_cryption
+[ -n "$(installer_get "DI_CRYPT_PASSWD")" ] && clean_disk_cryption
 clean_check_mode
 reboot
